[FIX] end token not returned any more
Before this fix, the model included the end token when returning the
resulting sequence. This is redundant, since the end of the sequence is
implicitly indicated by the end of the list anyway.

Additionally this messes up the BLEU score as well. Since the token is
returned, it also is considered in the calculation of the score.